requirements.txt
torch
transformers
datasets
pandas
tqdm
matplotlib
wandb
pyyaml
plotly
scikit-learn
nltk
simpletransformers